seperate it out
