+++ .env.example
@@ -1,5 +1,5 @@
 # 'mock' for generating mock proofs locally, 'local' for generating proofs locally, 'network' for generating proofs using the proving network.
-SP1_PROVER=mock
+SP1_PROVER=local
 # If using the proving network, set to your whitelisted private key. For more information, see:
 # https://docs.succinct.xyz/prover-network/setup.html#key-setup
 SP1_PRIVATE_KEY=
