+++ README.md
@@ -8,36 +8,70 @@ that can generate a proof of any RISC-V program.
 - [Rust](https://rustup.rs/)
 - [SP1](https://succinctlabs.github.io/sp1/getting-started/install.html)
 
-## Standard Proof Generation
+## Running the Project
 
-> [!WARNING]
-> You will need at least 16GB RAM to generate the default proof.
+There are three main ways to run this project: execute a program, generate a core proof, and
+generate an EVM-compatible proof.
 
-Generate the proof for your program using the standard prover.
+### Execute the Program
+
+To run the program without generating a proof:
 
 ```sh
 cd script
-RUST_LOG=info cargo run --bin prove --release
+cargo run --release -- --execute
 ```
 
-## EVM-Compatible Proof Generation & Verification
+This will execute the program and display the output.
+
+### Generate a Core Proof
+
+To generate a core proof for your program:
+
+```sh
+cd script
+cargo run --release -- --prove
+```
+
+### Generate an EVM-Compatible (PLONK)Proof
 
 > [!WARNING]
 > You will need at least 128GB RAM to generate the PLONK proof.
 
-Generate the proof that is small enough to be verified on-chain and verifiable by the EVM. This command also generates a fixture that can be used to test the verification of SP1 zkVM proofs inside Solidity.
+To generate a PLONK proof that is small enough to be verified on-chain and verifiable by the EVM:
 
 ```sh
 cd script
-RUST_LOG=info cargo run --bin prove --release -- --evm
+cargo run --release --bin evm
+```
+
+This command also generates a fixture that can be used to test the verification of SP1 zkVM proofs
+inside Solidity.
+
+### Retrieve the Verification Key
+
+To retrieve your `programVKey` for your on-chain contract, run the following command:
+
+```sh
+cargo run --release --bin vkey
 ```
 
 ## Using the Prover Network
 
-Make a copy of the example environment file:
+We highly recommend using the Succinct prover network for any non-trivial programs or benchmarking purposes. For more information, see the [setup guide](https://docs.succinct.xyz/prover-network/setup.html).
+
+To get started, copy the example environment file:
 
 ```sh
 cp .env.example .env
 ```
 
-Then, set the `SP1_PROVER` environment variable to `network` and set the `SP1_PRIVATE_KEY` environment variable to your whitelisted private key. For more information, see the [setup guide](https://docs.succinct.xyz/prover-network/setup.html).
+Then, set the `SP1_PROVER` environment variable to `network` and set the `SP1_PRIVATE_KEY`
+environment variable to your whitelisted private key.
+
+For example, to generate an EVM-compatible proof using the prover network, run the following
+command:
+
+```sh
+SP1_PROVER=network SP1_PRIVATE_KEY=... cargo run --release --bin evm
+```
