+++ program/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2021"
 
 [dependencies]
 alloy-sol-types = "0.7.2"
-sp1-zkvm = "1.0.1"
+sp1-zkvm = "1.1.0"
