+++ script/Cargo.toml
@@ -3,18 +3,22 @@
 version = "0.1.0"
 name = "fibonacci-script"
 edition = "2021"
-default-run = "prove"
+default-run = "fibonacci"
 
 [[bin]]
-name = "prove"
-path = "src/bin/prove.rs"
+name = "fibonacci"
+path = "src/bin/main.rs"
+
+[[bin]]
+name = "evm"
+path = "src/bin/evm.rs"
 
 [[bin]]
 name = "vkey"
 path = "src/bin/vkey.rs"
 
 [dependencies]
-sp1-sdk = "1.0.1"
+sp1-sdk = "1.1.0"
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 clap = { version = "4.0", features = ["derive", "env"] }
@@ -23,4 +27,4 @@ alloy-sol-types = "0.7.2"
 hex = "0.4.3"
 
 [build-dependencies]
-sp1-helper = "1.0.1"
+sp1-helper = "1.1.0"
