+++ script/src/bin/evm.rs
@@ -1,0 +1,93 @@
+use alloy_sol_types::{sol, SolType};
+use clap::Parser;
+use serde::{Deserialize, Serialize};
+use sp1_sdk::{HashableKey, ProverClient, SP1ProofWithPublicValues, SP1Stdin, SP1VerifyingKey};
+use std::path::PathBuf;
+
+/// The ELF (executable and linkable format) file for the Succinct RISC-V zkVM.
+pub const FIBONACCI_ELF: &[u8] = include_bytes!("../../../program/elf/riscv32im-succinct-zkvm-elf");
+
+/// The arguments for the EVM command.
+#[derive(Parser, Debug)]
+#[clap(author, version, about, long_about = None)]
+struct EVMArgs {
+    #[clap(long, default_value = "20")]
+    n: u32,
+}
+
+/// The public values encoded as a tuple that can be easily deserialized inside Solidity.
+type PublicValuesTuple = sol! {
+    tuple(uint32, uint32, uint32)
+};
+
+/// A fixture that can be used to test the verification of SP1 zkVM proofs inside Solidity.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct SP1FibonacciProofFixture {
+    a: u32,
+    b: u32,
+    n: u32,
+    vkey: String,
+    public_values: String,
+    proof: String,
+}
+
+fn main() {
+    // Setup the logger.
+    sp1_sdk::utils::setup_logger();
+
+    // Parse the command line arguments.
+    let args = EVMArgs::parse();
+
+    // Setup the prover client.
+    let client = ProverClient::new();
+
+    // Setup the program.
+    let (pk, vk) = client.setup(FIBONACCI_ELF);
+
+    // Setup the inputs.
+    let mut stdin = SP1Stdin::new();
+    stdin.write(&args.n);
+
+    println!("n: {}", args.n);
+
+    // Generate the proof.
+    let proof = client
+        .prove(&pk, stdin)
+        .plonk()
+        .run()
+        .expect("failed to generate proof");
+
+    create_plonk_fixture(&proof, &vk);
+}
+
+/// Create a fixture for the given proof.
+fn create_plonk_fixture(proof: &SP1ProofWithPublicValues, vk: &SP1VerifyingKey) {
+    // Deserialize the public values.
+    let bytes = proof.public_values.as_slice();
+    let (n, a, b) = PublicValuesTuple::abi_decode(bytes, false).unwrap();
+
+    // Create the testing fixture so we can test things end-to-end.
+    let fixture = SP1FibonacciProofFixture {
+        a,
+        b,
+        n,
+        vkey: vk.bytes32().to_string(),
+        public_values: format!("0x{}", hex::encode(bytes)),
+        proof: format!("0x{}", hex::encode(proof.bytes())),
+    };
+
+    // Print fixture information
+    println!("Verification Key: {}", fixture.vkey);
+    println!("Public Values: {}", fixture.public_values);
+    println!("Proof Bytes: {}", fixture.proof);
+
+    // Save the fixture to a file.
+    let fixture_path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../contracts/src/fixtures");
+    std::fs::create_dir_all(&fixture_path).expect("failed to create fixture path");
+    std::fs::write(
+        fixture_path.join("fixture.json"),
+        serde_json::to_string_pretty(&fixture).unwrap(),
+    )
+    .expect("failed to write fixture");
+}
